hw_13/bot/requirements.txt
python-telegram-bot>=21.0
transformers>=4.36.0
peft>=0.7.0
torch>=2.1.0
accelerate>=0.25.0
python-dotenv>=1.0.0
